fix error in playerstats due to trying to read pkl as json

+++ wrapper/api/minecraft.py
@@ -198,14 +198,17 @@ class Minecraft(object):
                 if player_uuid == self.wrapper.uuids.getuuidfromname(username):
                     continue
 
-            with open("wrapper-data/players/%s" % uuid_file_found) as f:
-                data = f.read()
-            try:
-                players[player_uuid] = json.loads(data, self._encoding)
-            except Exception as e:
-                self.log.error("Failed to load player data"
-                               " '%s':\n%s", player_uuid, e)
-                os.remove("wrapper-data/players/" + uuid_file_found)
+            # added because some files under other versions were pickling the data
+            if uuid_file_found[-4:] == "json":
+                with open("wrapper-data/players/%s" % uuid_file_found) as f:
+                    data = f.read()
+                try:
+                    players[player_uuid] = json.loads(data, self._encoding)
+                except Exception as e:
+                    self.log.error("Failed to load player data"
+                                   " '%s':\n%s", player_uuid, e)
+                    os.remove("wrapper-data/players/" + uuid_file_found)
+
         return players
 
     def getPlayers(self):  # returns a list of players
